chore: Update post-pr-merge workflow
Ignore the egg-info and dist directories in the .gitignore file. Also, update the post-pr-merge workflow to run only one at a time per branch and cancel if a run is in progress. Add the '.github/workflows/post-pr-merge.yaml' file to the list of paths triggering the workflow. Attest the build provenance in the workflow by using the 'actions/attest-build-provenance' action.

Refactor the code to improve the workflow and ensure better build management.

Signed-off-by: David Hooton <david@hooton.org>